Move math classes to common.
