+++ core/src/commonMain/kotlin/info/laht/threekt/math/Euler.kt
@@ -213,9 +213,7 @@ class Euler(
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as Euler
+        if (other !is Euler) return false
 
         if (_x != other._x) return false
         if (_y != other._y) return false
