+++ core/src/commonMain/kotlin/info/laht/threekt/math/Flattable.kt
