+++ core/src/commonMain/kotlin/info/laht/threekt/math/Matrix3.kt
@@ -1,9 +1,6 @@
 package info.laht.threekt.math
 
 import info.laht.threekt.core.Cloneable
-import info.laht.threekt.core.FloatBufferAttribute
-import org.lwjgl.BufferUtils
-import java.nio.FloatBuffer
 import kotlin.math.cos
 import kotlin.math.sin
 
@@ -66,23 +63,6 @@ class Matrix3(
         )
 
         return this
-    }
-
-    fun applyToBufferAttribute(attribute: FloatBufferAttribute): FloatBufferAttribute {
-        val v1 = Vector3()
-
-        for (i in 0 until attribute.count) {
-            v1.x = attribute.getX(i)
-            v1.y = attribute.getY(i)
-            v1.z = attribute.getZ(i)
-
-            v1.applyMatrix3(this)
-
-            attribute.setXYZ(i, v1.x, v1.y, v1.z)
-        }
-
-        return attribute
-
     }
 
     /**
@@ -327,16 +307,6 @@ class Matrix3(
 
     override fun toArray(array: FloatArray?, offset: Int): FloatArray {
         return elements.copyInto(array ?: FloatArray(9), offset)
-    }
-
-    fun toBuffer(buffer: FloatBuffer?, offset: Int): FloatBuffer {
-
-        val buf = buffer ?: BufferUtils.createFloatBuffer(size)
-        elements.forEachIndexed { i, v ->
-            buf.put(i + offset, v)
-        }
-
-        return buf
     }
 
     override fun equals(other: Any?): Boolean {
