+++ core/src/commonMain/kotlin/info/laht/threekt/math/Plane.kt
@@ -1,6 +1,7 @@
 package info.laht.threekt.math
 
 import info.laht.threekt.core.Cloneable
+import kotlin.jvm.JvmOverloads
 
 
 class Plane @JvmOverloads constructor(
