+++ core/src/commonMain/kotlin/info/laht/threekt/math/Quaternion.kt
@@ -1,6 +1,7 @@
 package info.laht.threekt.math
 
 import info.laht.threekt.core.Cloneable
+import kotlin.jvm.JvmOverloads
 import kotlin.math.*
 
 class Quaternion(
@@ -477,9 +478,7 @@ class Quaternion(
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as Quaternion
+        if (other !is Quaternion) return false
 
         if (_x != other._x) return false
         if (_y != other._y) return false
