+++ core/src/commonMain/kotlin/info/laht/threekt/math/Ray.kt
@@ -1,5 +1,7 @@
 package info.laht.threekt.math
 
+import kotlin.jvm.JvmOverloads
+
 class Ray @JvmOverloads constructor(
     var origin: Vector3 = Vector3(),
     val direction: Vector3 = Vector3()
