+++ core/src/commonMain/kotlin/info/laht/threekt/math/Sphere.kt
@@ -1,6 +1,7 @@
 package info.laht.threekt.math
 
 import info.laht.threekt.core.Cloneable
+import kotlin.jvm.JvmOverloads
 import kotlin.math.abs
 import kotlin.math.max
 import kotlin.math.sqrt
@@ -117,9 +118,7 @@ class Sphere(
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as Sphere
+        if (other !is Sphere) return false
 
         if (center != other.center) return false
         if (radius != other.radius) return false
