+++ core/src/commonMain/kotlin/info/laht/threekt/math/SphericalHarmonics3.kt
