+++ core/src/jvmMain/kotlin/info/laht/threekt/objects/Line.kt
@@ -3,6 +3,7 @@ package info.laht.threekt.objects
 import info.laht.threekt.core.*
 import info.laht.threekt.materials.LineBasicMaterial
 import info.laht.threekt.math.Vector3
+import info.laht.threekt.math.fromBufferAttribute
 
 open class Line @JvmOverloads constructor(
